# control-plane/internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: Agent Sandbox Platform control-plane API
  # The control plane serves this document at GET /openapi.json with the version of its own build here.
  version: dev
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  description: |
    The HTTP API of the Agent Sandbox Platform control plane: it creates sandboxes (microVMs) on
    the nodes, runs commands in them, and manages the nodes, API keys and egress policy.

    This document is written by hand and checked against the code: a test fails when a route of the
    control plane is missing here (or the other way round), when a field of a JSON body differs
    from the Go type that carries it, or when a handler can answer a status that is not listed.
    The reference page `docs/reference/api.md` is generated from it (`make docs`). While ASP is
    0.x, a minor release may change the API; [CHANGELOG.md](https://github.com/luisgf/agent-sandbox-platform/blob/main/CHANGELOG.md) says so.

    ## Authentication

    Every route except the public ones needs `Authorization: Bearer <token>`. A token is one of:

    * an **API key** (`asp_` and 40 hexadecimal digits), made with `POST /v1/api-keys`. A *tenant*
      key acts only within its tenant. A *platform* key sees every tenant and may call the node
      routes, which a tenant key never can.
    * an **IdP access token**: a JWT of the OIDC provider the control plane trusts
...

...

...

    ### Roles

...

    ### Tenants

...

...

...

...

...

...

...
servers:
...
security:
  - apiKey: []
  - idpToken: []
...
paths:
...
  /v1/sandboxes/{id}:
    get:
      operationId: getSandbox
      summary: Get a sandbox
      description: The sandbox and the state its node last reported. An IdP caller needs to be its owner or an `admin`, `operator` or `viewer`.
      tags: [Sandboxes]
      x-asp-access: tenant
      parameters:
        - $ref: '#/components/parameters/SandboxID'
      responses:
        '200':
          description: The sandbox.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sandbox'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      operationId: deleteSandbox
      summary: Delete a sandbox and its disk
      description: |
        Deletes the VM and its disk. The sandbox goes to `deleting`, and to `deleted` when its node
        reports it removed (at once when no node holds anything). The row stays for the audit trail.
        Deleting a sandbox that is being or already deleted changes nothing and answers 200.
        An IdP caller needs to be the owner, an `admin`, or an `operator` with `destroy-any`.
      tags: [Sandboxes]
      x-asp-access: tenant
      parameters:
        - $ref: '#/components/parameters/SandboxID'
      responses:
        '200':
          description: The sandbox, `deleting` or `deleted`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sandbox'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /v1/sandboxes/{id}/attestation:
    get:
      operationId: getAttestation
      summary: The boot evidence of a sandbox
      description: |
        The latest signed statement a node made about what it booted the sandbox from, with whether it is
        `fresh`, whether the node `measured` the kernel and image (hashed them) and, when the control plane has an
        image allowlist, whether the digests are on it (`allowlisted`, `image_name`). The statement is the node's
        own word, signed by it: a node the control plane trusts, not a hardware root of trust.
      tags: [Attestation]
      x-asp-access: tenant
      parameters:
        - $ref: '#/components/parameters/SandboxID'
      responses:
        '200':
          description: The evidence.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttestationResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: The sandbox does not exist, or it has no attestation yet.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/attestation/verify:
    post:
      operationId: verifyAttestation
      summary: Verify boot evidence without storing it
      description: Checks the signature and shape of an evidence bundle and says whether it is fresh. It stores nothing and looks at no sandbox.
      tags: [Attestation]
      x-asp-access: tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyAttestationRequest'
      responses:
        '200':
          description: The verdict. `valid` false carries the reason in `error`; it is still a 200.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyAttestationResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '503':
          $ref: '#/components/responses/Unavailable'
  /v1/sandboxes/{id}/attest:
    post:
      operationId: storeAttestation
      summary: A node submits the boot evidence of a sandbox
      description: |
        The node-agent sends the statement it signed when it booted a sandbox. The control plane verifies the
        signature against a key it trusts (`ASP_ATTEST_KEY`, `ASP_ATTEST_PUB`, `ASP_ATTEST_TRUSTED_PUBS`) or, over
        mTLS, the key of the node's certificate, which binds the evidence to the node's enrolled identity. A key that
        only the statement carries is not trusted. It then checks the image against the allowlist if there is one,
        stores the evidence and records a `sandbox.attested` event; a refused image is recorded as
        `sandbox.attestation_refused`.
      tags: [Node agent]
      x-asp-access: node
      security:
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/SandboxID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttestRequest'
      responses:
        '200':
          description: The stored evidence.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttestationRecord'
        '400':
          description: The statement is for another sandbox, its signature does not verify, or its image is not allowed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: |
        An API key: `asp_` and 40 hexadecimal digits. The secret is shown once, when the key is created or
        rotated (`POST /v1/api-keys`); the control plane keeps only a hash of it. A *tenant* key acts within its
        tenant; a *platform* key sees every tenant and may call the node routes.
    idpToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        An access token of the OIDC identity provider the control plane trusts (`ASP_IDP_ISSUER`). Its claims give the
        tenant, the subject, the e-mail and the role (`admin`, `operator`, `user`, `viewer`).
    enrollToken:
      type: http
      scheme: bearer
      description: |
        The node bootstrap token (`ASP_NODE_BOOTSTRAP_TOKEN`) or a single-use enrollment token (`asp_enroll_…`, from
        `POST /v1/nodes/enroll-tokens`). `X-ASP-Bootstrap-Token` carries it too.
  parameters:
    SandboxID:
      name: id
      in: path
      required: true
      description: The id of the sandbox.
      schema:
        type: string
...
  responses:
    BadRequest:
      description: The body is not valid JSON, a field is missing or out of range, or a parameter is invalid. The message says which.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: No credential, an unknown, revoked or expired one, or an IdP token that is not valid or names no tenant.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: The credential is valid but its role, its tenant or its node identity does not allow this.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: It does not exist, or it belongs to another tenant than the caller's.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    PayloadTooLarge:
      description: The body is larger than the limit (1 MiB; 64 KiB for the calls of nodes and keys).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: The control plane failed to do it, usually its store. The message is the error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Unavailable:
      description: The control plane has not been set up for this (attestation, OIDC, enrollment CA).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      description: The body of every error.
      required: [error]
      properties:
        error:
          type: string
          description: What went wrong, for people. It may change between releases.
...
    SandboxState:
      type: string
      description: Where a sandbox is in its life. See the states in the description of this API.
      enum: [requested, scheduled, starting, running, paused, stopping, stopped, failed, deleting, deleted]
    Sandbox:
      type: object
      description: A sandbox, as the control plane and its node last knew it.
      required:
        - id
        - tenant_id
        - node_id
        - state
        - vmm_profile
        - image_ref
        - cpu_millis
        - memory_mib
        - state_version
        - last_activity_at
        - boot_count
        - local_net
        - local_net_state
        - created_at
        - updated_at
      properties:
        id:
          description: 'The id of the sandbox.'
          type: string
        tenant_id:
          description: 'The tenant that owns it.'
          type: string
        node_id:
          type: string
          nullable: true
          description: The node the sandbox is placed on; null if it has none.
        state:
          $ref: '#/components/schemas/SandboxState'
        vmm_profile:
          type: string
          description: The hypervisor profile; `cloud-hypervisor` unless asked otherwise.
        image_ref:
          type: string
          description: The image the caller asked for, recorded with the sandbox.
        cpu_millis:
          type: integer
          description: CPU in thousandths of a core.
        memory_mib:
          description: 'Memory in MiB.'
          type: integer
        state_version:
          type: integer
          format: int64
          description: Goes up with every change of state.
        owner_sub:
          type: string
          description: The IdP subject of the creator. Empty for a sandbox made with an API key.
        owner_email:
          description: 'The creator''s e-mail, from the IdP token. For showing; it decides nothing.'
          type: string
        last_activity_at:
          type: string
          format: date-time
          description: The last command, or the last time the sandbox started. The idle reaper counts from here.
        stop_reason:
          type: string
          description: |
            Why the sandbox was stopped or failed, when the user did not ask: `idle_timeout`, `node_lost`,
            `node_agent_restarted`, `vmm_exited`, `unscheduled`, `retention_expired` or `tenant_cap`.
        status_detail:
          type: string
          description: Why the node last reported a failure, or why a resume went back to stopped.
        boot_count:
          type: integer
          description: >-
            How many times the sandbox has started: 1 at the first boot, one more per resume.
        booted_at:
          type: string
          format: date-time
          description: When the node first reported it running. Until then there is no disk worth keeping. Never cleared.
        stopped_at:
          type: string
          format: date-time
          description: When the node reported it stopped; absent while it is not. The retention period counts from it.
        workspace_host_path:
          type: string
          description: The host directory shared into the guest at `/workspace`, if any.
        local_net:
          type: boolean
          description: The sandbox routes its traffic through the caller's local-net tunnel and never through the node's public proxy.
        local_net_state:
          description: 'Where the tunnel is: `off` (none; egress goes through the node''s proxy), `pending` (asked for, not up), `up`, or `withdrawn` (taken down; egress stays closed).'
          type: string
          enum: ['off', pending, up, withdrawn]
        local_net_attached_at:
          description: 'When the tunnel last came up.'
          type: string
          format: date-time
        local_net_grant_expires_at:
          description: 'When the current grant expires. The grant itself is not stored.'
          type: string
          format: date-time
        local_net_client_public:
          type: string
          description: The WireGuard public key of the caller's end.
        local_net_node_public:
          type: string
          description: The WireGuard public key of the node's end.
        local_net_listen_port:
          description: 'The node''s UDP port for the tunnel; 0 until the node publishes it.'
          type: integer
        local_net_node_addr:
          description: 'The node''s end of the tunnel''s /30.'
          type: string
        local_net_client_addr:
          description: 'The caller''s end of the tunnel''s /30.'
          type: string
        created_at:
          description: 'When the sandbox was created.'
          type: string
          format: date-time
        updated_at:
          description: 'When it last changed.'
          type: string
          format: date-time
...
    BootStatement:
      type: object
      description: What a node says it booted a sandbox from.
      required: [sandbox_id, image_digest, vmm_profile, cid, node_id, ts]
      properties:
        sandbox_id:
          description: 'The sandbox the VM was booted for.'
          type: string
        image_digest:
          type: string
          description: '`sha256:<hex>` of the base image the sandbox''s disk was copied from, as the node hashed it. A node that does not measure sends the image reference here.'
        vmm_profile:
          description: 'The hypervisor profile it ran under.'
          type: string
        cid:
          type: integer
          minimum: 0
          description: The vsock context id of the guest.
        node_id:
          description: 'The node that booted it.'
          type: string
        ts:
          type: string
          format: date-time
          description: When the node made the statement. Old statements are not fresh.
        kernel_digest:
          type: string
          description: '`sha256:<hex>` of the kernel the VM loaded.'
        vmm_version:
          type: string
          description: The version the hypervisor reported.
        boot:
          type: string
          enum: [new, resume]
          description: A `resume` boots a retained disk that has diverged from the base image `image_digest` names.
    Evidence:
      type: object
      description: A signed boot statement.
      required: [statement, signature, alg]
      properties:
        statement:
          $ref: '#/components/schemas/BootStatement'
        signature:
          type: string
          description: base64url of the DER ECDSA signature.
        alg:
          description: 'The signature algorithm.'
          type: string
          enum: [ES256]
        key_id:
          description: 'The id of the key that signed, when the node has one.'
          type: string
        public_key_pem:
          description: 'The public key, when the node sends it along.'
          type: string
        received_at:
          description: 'When the control plane stored it.'
          type: string
          format: date-time
    AttestRequest:
      type: object
      required: [statement, signature]
      properties:
        statement:
          $ref: '#/components/schemas/BootStatement'
        signature:
          description: 'base64url of the DER ECDSA signature of the statement.'
          type: string
        alg:
          type: string
          enum: [ES256]
          description: Defaults to `ES256`.
        key_id:
          description: 'The id of the signing key.'
          type: string
        public_key_pem:
          description: 'The signing key, PEM.'
          type: string
    AttestationRecord:
      type: object
      description: The latest verified evidence kept for a sandbox.
      required: [sandbox_id, node_id, image_digest, vmm_profile, cid, statement_ts, alg, signature, bundle, received_at]
      properties:
        sandbox_id:
          description: 'The sandbox.'
          type: string
        node_id:
          description: 'The node that booted it.'
          type: string
        image_digest:
          description: '`sha256:<hex>` of the base image, or the image reference for a node that does not measure.'
          type: string
        vmm_profile:
          description: 'The hypervisor profile.'
          type: string
        cid:
          description: 'The vsock context id of the guest.'
          type: integer
          minimum: 0
        statement_ts:
          description: 'When the node made the statement.'
          type: string
          format: date-time
        alg:
          description: 'The signature algorithm.'
          type: string
        key_id:
          description: 'The id of the signing key.'
          type: string
        signature:
          description: 'base64url of the DER ECDSA signature.'
          type: string
        bundle:
          $ref: '#/components/schemas/Evidence'
        received_at:
          description: 'When the control plane stored it.'
          type: string
          format: date-time
    AttestationResult:
      type: object
      required: [attestation, fresh, measured]
      properties:
        attestation:
          $ref: '#/components/schemas/AttestationRecord'
        fresh:
          type: boolean
          description: The statement is recent enough to be trusted for a token claim.
        measured:
          type: boolean
          description: The statement carries real digests of the kernel and the base image.
        allowlisted:
          type: boolean
          description: >-
            Only when the control plane has an image allowlist: the digests are on it.
        image_name:
          type: string
          description: The allowlist's name for the image, when `allowlisted`.
    VerifyAttestationRequest:
      type: object
      required: [evidence]
      properties:
        evidence:
          $ref: '#/components/schemas/Evidence'
    VerifyAttestationResult:
      type: object
      required: [valid, error, fresh, kid, alg]
      properties:
        valid:
          description: 'Whether the signature and the statement hold.'
          type: boolean
        error:
          type: string
          description: Why it is not valid; empty when it is.
        fresh:
          description: 'Whether the statement is recent enough.'
          type: boolean
        kid:
          description: 'The key id of the evidence.'
          type: string
        alg:
          description: 'Its algorithm.'
          type: string
...